refactor(telemetry): update HandlerFactory function to use telemetry package.
